Add ConvertBooleanToYesNoFilter
Move ConvertBooleanToYesNoFilter from CensusDataFilter to specific filter.
