+++ src/AppBundle/Util/Filter/CensusDataFilter.php
@@ -10,16 +10,7 @@ class CensusDataFilter extends AbstractExtension
 {
     
 
-    public function convertBooleanToYesNoFilter($answer)
-    {
-        if ($answer === true) {
-            return "<span style='color: darkgreen'>Sim</span>";
-        } elseif ($answer === false) {
-            return "<span style='color: red'>Não</span>";
-        }
-
-        return "<span style='color: #666'>Não informado</span>";
-    }
+    
 
     public function waterSupplierFilter($censusServices)
     {
@@ -41,7 +32,6 @@ class CensusDataFilter extends AbstractExtension
     public function getFilters()
     {
         return [
-            new TwigFilter('convertNumberToYesNoFilter', [$this, 'convertNumberToYesNoFilter']),
             new TwigFilter('waterSupplierFilter', [$this, 'waterSupplierFilter']),
         ];
     }
